+++ src/editor/source-browser-editor.ts
@@ -4,6 +4,7 @@ import { css, html, TemplateResult } from 'lit';
 // our imports.
 import { BaseEditor } from './base-editor';
 import { Section } from '../types/section';
+import { EDITOR_DEFAULT_BROWSER_ITEMS_PER_ROW } from '../constants';
 
 
 const CONFIG_SETTINGS_SCHEMA = [
@@ -27,7 +28,7 @@ const CONFIG_SETTINGS_SCHEMA = [
     help: 'use 1 for list format',
     required: true,
     type: 'integer',
-    default: 3,
+    default: EDITOR_DEFAULT_BROWSER_ITEMS_PER_ROW,
     valueMin: 1,
     valueMax: 12,
   },